✂️style: BoxOfficeDecoder를 기능분리해주고 네이밍 BoxOfficeManager로 변경
